Backend/cardmaster-assistant/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
boto3==1.26.151
botocore==1.29.151
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.3
cohere==4.26.1
EasyProcess==1.1
entrypoint2==1.1
fastavro==1.8.2
firebase==4.0.1
Flask==2.3.2
frozenlist==1.4.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
nulltype==2.3.1
Pillow==10.0.0
plaid-python==14.0.0
psycopg==3.1.10
psycopg-binary==3.1.10
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
requests==2.31.0
rfc3986==1.5.0
rubicon-objc==0.4.6
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
